% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutputTableFromKaKsCalc.R
\name{createOutputTableFromKaKsCalc}
\alias{createOutputTableFromKaKsCalc}
\title{Create output table from KaKs_caclculator data}
\usage{
createOutputTableFromKaKsCalc(directoryPath, outfileName,
  extension = NULL)
}
\arguments{
\item{directoryPath, }{is the path to the directory containing output from KaKs_Calculator}

\item{extension, }{the extension of all the output files from KaKs_Calculator}
}
\description{
Create output table from KaKs_caclculator data
}
