% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaKsMap.R
\name{KaKsMap}
\alias{KaKsMap}
\title{KaKs Map}
\usage{
KaKsMap(treePath, pathToDirWithKaKsCalcData, extensionForKaKsData = NULL,
  fileNameKaKsTable, mainCircleRadius = 0.5, radiusGeneDots = 0.02)
}
\arguments{
\item{treePath}{path to the newick format tree}

\item{pathToDirWithKaKsCalcData}{path to a directory containing output files from KaKs_Calculator}

\item{extensionForKaKsData}{if there are other files in the same dir provided in pathToDirWithKaKsCalcData indicate
the desired files using the extention}

\item{fileNameKaKsTable}{file name to write the compiled data to for future reference}

\item{mainCircleRadius}{desired radius for the main circle}

\item{radiusGeneDots}{desired radius for the genes which are plotted as dots around the circumfrence of the main circle}
}
\description{
KaKs Map
}
