% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesOnCircle.R
\name{plotGenesOnCircle}
\alias{plotGenesOnCircle}
\title{Plot genes on the main circle}
\usage{
plotGenesOnCircle(tree, radius, x = 0, y = 0, clustered = FALSE,
  geneCircleRadius, plot = plot.new())
}
\arguments{
\item{y}{coordinate for the center of the circle}

\item{clustered}{default FALSE results in evenly spaced genes on the circle, functionality for TRUE unimplemented}

\item{geneCircleRadius}{radius for genes placed around the edge of the circle}

\item{plot}{default is to create a new plot}
}
\value{
return the coordinates for the gene placement on the circumfrence of the main circle
}
\description{
Plot genes on the main circle
}
