% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinesOfKaKsValues.R
\name{plotLinesOfKaKsValues}
\alias{plotLinesOfKaKsValues}
\title{Plot lines of KaKsValues}
\usage{
plotLinesOfKaKsValues(tree, table, points, plot = plot.new())
}
\arguments{
\item{tree}{phylo object created by using read.tree on a newick tree format}

\item{table}{containing output from KaKs_Calculator}

\item{points}{points where the genes were placed around the circumfrence of the main circle in plotGenesOnCircle}

\item{plot}{default is to create a new plot}
}
\description{
Plot lines of KaKsValues
}
