% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInData.R
\name{readInData}
\alias{readInData}
\title{Read in data}
\usage{
readInData(phylogeneticTreePath, kaksCalcOutputTablePath)
}
\arguments{
\item{phylogeneticTreePath}{path to a tree in newick format}

\item{kaksCalcOutput}{path to a table containing a header and output from KaKs_Calculator for at least one calculation}
}
\value{
a list containing a phylo object and a table
}
\description{
Read in data
}
